requirements.txt
boto3>=1.34.0
requests>=2.31.0
requests-aws4auth>=1.2.3
opensearch-py>=2.4.0
python-dotenv>=1.0.0
streamlit>=1.35.0
anthropic>=0.8.0
pillow>=10.0.0
fitz>=0.0.1.dev2
PyMuPDF>=1.22.0 